requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
datasets==2.20.0
dill==0.3.8
docker-pycreds==0.4.0
donut-python==1.0.9
et-xmlfile==1.1.0
exceptiongroup==1.2.1
fastapi==0.96.0
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.64.1
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.23.4
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
lightning-utilities==0.11.2
Markdown==3.6
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
munch==4.0.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
openpyxl==3.1.4
packaging==24.1
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.2
protobuf==4.25.3
psutil==6.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==1.10.17
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==1.9.5
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.3
sconf==0.2.5
sentencepiece==0.2.0
sentry-sdk==2.6.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.27.0
sympy==1.12.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
timm==1.0.7
tk==0.1.0
tokenizers==0.19.1
torch==2.3.1
torchmetrics==1.4.0.post0
torchvision==0.18.1
tqdm==4.66.4
transformers @ git+https://github.com/huggingface/transformers.git@74b92c62560b7ade42d35a49f9063adc8b805c4a
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
wandb==0.17.2
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.3
xxhash==3.4.1
yarl==1.9.4
zss==1.2.0
